refactor: split tests into several files, create the `utils` directory
