+++ test/main.js
@@ -33,13 +33,13 @@ import HeadersOrig, {createHeadersLenient} from '../src/headers';
 import RequestOrig from '../src/request';
 import ResponseOrig from '../src/response';
 import Body, {getTotalBytes, extractContentType} from '../src/body';
-import TestServer from './server';
+import TestServer from './utils/server';
 
 const {
 	Uint8Array: VMUint8Array
 } = vm.runInNewContext('this');
 
-import chaiTimeout from './chai-timeout';
+import chaiTimeout from './utils/chai-timeout';
 
 chai.use(chaiPromised);
 chai.use(chaiIterator);
@@ -59,6 +59,18 @@ after(done => {
 });
 
 const itIf = val => val ? it : it.skip;
+
+function streamToPromise(stream, dataHandler) {
+	return new Promise((resolve, reject) => {
+		stream.on('data', (...args) => {
+			Promise.resolve()
+				.then(() => dataHandler(...args))
+				.catch(reject);
+		});
+		stream.on('end', resolve);
+		stream.on('error', reject);
+	});
+}
 
 describe('node-fetch', () => {
 	it('should return a promise', () => {
@@ -1080,8 +1092,8 @@ describe('node-fetch', () => {
 					try {
 						expect(error).to.be.an.instanceof(Error).and.have.property('name', 'AbortError');
 						resolve();
-					} catch (error2) {
-						reject(error2);
+					} catch (error_) {
+						reject(error_);
 					}
 				});
 			}),
@@ -1366,7 +1378,7 @@ describe('node-fetch', () => {
 
 	itIf(process.platform !== 'win32')('should allow POST request with form-data using stream as body', () => {
 		const form = new FormData();
-		form.append('my_field', fs.createReadStream(path.join(__dirname, 'dummy.txt')));
+		form.append('my_field', fs.createReadStream(path.join(__dirname, './utils/dummy.txt')));
 
 		const url = `${base}multipart`;
 		const opts = {
@@ -2144,717 +2156,5 @@ describe('node-fetch', () => {
 		expect(extractContentType(bufferBody)).to.be.null;
 		expect(extractContentType(bodyContent)).to.equal('text/plain;charset=UTF-8');
 		expect(extractContentType(null)).to.be.null;
-	});
-});
-
-describe('Headers', () => {
-	it('should have attributes conforming to Web IDL', () => {
-		const headers = new Headers();
-		expect(Object.getOwnPropertyNames(headers)).to.be.empty;
-		const enumerableProperties = [];
-
-		for (const property in headers) {
-			enumerableProperties.push(property);
-		}
-
-		for (const toCheck of [
-			'append',
-			'delete',
-			'entries',
-			'forEach',
-			'get',
-			'has',
-			'keys',
-			'set',
-			'values'
-		]) {
-			expect(enumerableProperties).to.contain(toCheck);
-		}
-	});
-
-	it('should allow iterating through all headers with forEach', () => {
-		const headers = new Headers([
-			['b', '2'],
-			['c', '4'],
-			['b', '3'],
-			['a', '1']
-		]);
-		expect(headers).to.have.property('forEach');
-
-		const result = [];
-		headers.forEach((val, key) => {
-			result.push([key, val]);
-		});
-
-		expect(result).to.deep.equal([
-			['a', '1'],
-			['b', '2, 3'],
-			['c', '4']
-		]);
-	});
-
-	it('should allow iterating through all headers with for-of loop', () => {
-		const headers = new Headers([
-			['b', '2'],
-			['c', '4'],
-			['a', '1']
-		]);
-		headers.append('b', '3');
-		expect(headers).to.be.iterable;
-
-		const result = [];
-		for (const pair of headers) {
-			result.push(pair);
-		}
-
-		expect(result).to.deep.equal([
-			['a', '1'],
-			['b', '2, 3'],
-			['c', '4']
-		]);
-	});
-
-	it('should allow iterating through all headers with entries()', () => {
-		const headers = new Headers([
-			['b', '2'],
-			['c', '4'],
-			['a', '1']
-		]);
-		headers.append('b', '3');
-
-		expect(headers.entries()).to.be.iterable
-			.and.to.deep.iterate.over([
-				['a', '1'],
-				['b', '2, 3'],
-				['c', '4']
-			]);
-	});
-
-	it('should allow iterating through all headers with keys()', () => {
-		const headers = new Headers([
-			['b', '2'],
-			['c', '4'],
-			['a', '1']
-		]);
-		headers.append('b', '3');
-
-		expect(headers.keys()).to.be.iterable
-			.and.to.iterate.over(['a', 'b', 'c']);
-	});
-
-	it('should allow iterating through all headers with values()', () => {
-		const headers = new Headers([
-			['b', '2'],
-			['c', '4'],
-			['a', '1']
-		]);
-		headers.append('b', '3');
-
-		expect(headers.values()).to.be.iterable
-			.and.to.iterate.over(['1', '2, 3', '4']);
-	});
-
-	it('should reject illegal header', () => {
-		const headers = new Headers();
-		expect(() => new Headers({'He y': 'ok'})).to.throw(TypeError);
-		expect(() => new Headers({'Hé-y': 'ok'})).to.throw(TypeError);
-		expect(() => new Headers({'He-y': 'ăk'})).to.throw(TypeError);
-		expect(() => headers.append('Hé-y', 'ok')).to.throw(TypeError);
-		expect(() => headers.delete('Hé-y')).to.throw(TypeError);
-		expect(() => headers.get('Hé-y')).to.throw(TypeError);
-		expect(() => headers.has('Hé-y')).to.throw(TypeError);
-		expect(() => headers.set('Hé-y', 'ok')).to.throw(TypeError);
-		// Should reject empty header
-		expect(() => headers.append('', 'ok')).to.throw(TypeError);
-
-		// 'o k' is valid value but invalid name
-		new Headers({'He-y': 'o k'});
-	});
-
-	it('should ignore unsupported attributes while reading headers', () => {
-		const FakeHeader = function () { };
-		// Prototypes are currently ignored
-		// This might change in the future: #181
-		FakeHeader.prototype.z = 'fake';
-
-		const res = new FakeHeader();
-		res.a = 'string';
-		res.b = ['1', '2'];
-		res.c = '';
-		res.d = [];
-		res.e = 1;
-		res.f = [1, 2];
-		res.g = {a: 1};
-		res.h = undefined;
-		res.i = null;
-		res.j = NaN;
-		res.k = true;
-		res.l = false;
-		res.m = Buffer.from('test');
-
-		const h1 = new Headers(res);
-		h1.set('n', [1, 2]);
-		h1.append('n', ['3', 4]);
-
-		const h1Raw = h1.raw();
-
-		expect(h1Raw.a).to.include('string');
-		expect(h1Raw.b).to.include('1,2');
-		expect(h1Raw.c).to.include('');
-		expect(h1Raw.d).to.include('');
-		expect(h1Raw.e).to.include('1');
-		expect(h1Raw.f).to.include('1,2');
-		expect(h1Raw.g).to.include('[object Object]');
-		expect(h1Raw.h).to.include('undefined');
-		expect(h1Raw.i).to.include('null');
-		expect(h1Raw.j).to.include('NaN');
-		expect(h1Raw.k).to.include('true');
-		expect(h1Raw.l).to.include('false');
-		expect(h1Raw.m).to.include('test');
-		expect(h1Raw.n).to.include('1,2');
-		expect(h1Raw.n).to.include('3,4');
-
-		expect(h1Raw.z).to.be.undefined;
-	});
-
-	it('should wrap headers', () => {
-		const h1 = new Headers({
-			a: '1'
-		});
-		const h1Raw = h1.raw();
-
-		const h2 = new Headers(h1);
-		h2.set('b', '1');
-		const h2Raw = h2.raw();
-
-		const h3 = new Headers(h2);
-		h3.append('a', '2');
-		const h3Raw = h3.raw();
-
-		expect(h1Raw.a).to.include('1');
-		expect(h1Raw.a).to.not.include('2');
-
-		expect(h2Raw.a).to.include('1');
-		expect(h2Raw.a).to.not.include('2');
-		expect(h2Raw.b).to.include('1');
-
-		expect(h3Raw.a).to.include('1');
-		expect(h3Raw.a).to.include('2');
-		expect(h3Raw.b).to.include('1');
-	});
-
-	it('should accept headers as an iterable of tuples', () => {
-		let headers;
-
-		headers = new Headers([
-			['a', '1'],
-			['b', '2'],
-			['a', '3']
-		]);
-		expect(headers.get('a')).to.equal('1, 3');
-		expect(headers.get('b')).to.equal('2');
-
-		headers = new Headers([
-			new Set(['a', '1']),
-			['b', '2'],
-			new Map([['a', null], ['3', null]]).keys()
-		]);
-		expect(headers.get('a')).to.equal('1, 3');
-		expect(headers.get('b')).to.equal('2');
-
-		headers = new Headers(new Map([
-			['a', '1'],
-			['b', '2']
-		]));
-		expect(headers.get('a')).to.equal('1');
-		expect(headers.get('b')).to.equal('2');
-	});
-
-	it('should throw a TypeError if non-tuple exists in a headers initializer', () => {
-		expect(() => new Headers([['b', '2', 'huh?']])).to.throw(TypeError);
-		expect(() => new Headers(['b2'])).to.throw(TypeError);
-		expect(() => new Headers('b2')).to.throw(TypeError);
-		expect(() => new Headers({[Symbol.iterator]: 42})).to.throw(TypeError);
-	});
-});
-
-describe('Response', () => {
-	it('should have attributes conforming to Web IDL', () => {
-		const res = new Response();
-		const enumerableProperties = [];
-		for (const property in res) {
-			enumerableProperties.push(property);
-		}
-
-		for (const toCheck of [
-			'body',
-			'bodyUsed',
-			'arrayBuffer',
-			'blob',
-			'json',
-			'text',
-			'url',
-			'status',
-			'ok',
-			'redirected',
-			'statusText',
-			'headers',
-			'clone'
-		]) {
-			expect(enumerableProperties).to.contain(toCheck);
-		}
-
-		for (const toCheck of [
-			'body',
-			'bodyUsed',
-			'url',
-			'status',
-			'ok',
-			'redirected',
-			'statusText',
-			'headers'
-		]) {
-			expect(() => {
-				res[toCheck] = 'abc';
-			}).to.throw();
-		}
-	});
-
-	it('should support empty options', () => {
-		let body = resumer().queue('a=1').end();
-		body = body.pipe(new stream.PassThrough());
-		const res = new Response(body);
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support parsing headers', () => {
-		const res = new Response(null, {
-			headers: {
-				a: '1'
-			}
-		});
-		expect(res.headers.get('a')).to.equal('1');
-	});
-
-	it('should support text() method', () => {
-		const res = new Response('a=1');
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support json() method', () => {
-		const res = new Response('{"a":1}');
-		return res.json().then(result => {
-			expect(result.a).to.equal(1);
-		});
-	});
-
-	it('should support buffer() method', () => {
-		const res = new Response('a=1');
-		return res.buffer().then(result => {
-			expect(result.toString()).to.equal('a=1');
-		});
-	});
-
-	it('should support blob() method', () => {
-		const res = new Response('a=1', {
-			method: 'POST',
-			headers: {
-				'Content-Type': 'text/plain'
-			}
-		});
-		return res.blob().then(result => {
-			expect(result).to.be.an.instanceOf(Blob);
-			expect(result.size).to.equal(3);
-			expect(result.type).to.equal('text/plain');
-		});
-	});
-
-	it('should support clone() method', () => {
-		let body = resumer().queue('a=1').end();
-		body = body.pipe(new stream.PassThrough());
-		const res = new Response(body, {
-			headers: {
-				a: '1'
-			},
-			url: base,
-			status: 346,
-			statusText: 'production'
-		});
-		const cl = res.clone();
-		expect(cl.headers.get('a')).to.equal('1');
-		expect(cl.url).to.equal(base);
-		expect(cl.status).to.equal(346);
-		expect(cl.statusText).to.equal('production');
-		expect(cl.ok).to.be.false;
-		// Clone body shouldn't be the same body
-		expect(cl.body).to.not.equal(body);
-		return cl.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support stream as body', () => {
-		let body = resumer().queue('a=1').end();
-		body = body.pipe(new stream.PassThrough());
-		const res = new Response(body);
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support string as body', () => {
-		const res = new Response('a=1');
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support buffer as body', () => {
-		const res = new Response(Buffer.from('a=1'));
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support ArrayBuffer as body', () => {
-		const res = new Response(stringToArrayBuffer('a=1'));
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support blob as body', () => {
-		const res = new Response(new Blob(['a=1']));
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support Uint8Array as body', () => {
-		const res = new Response(new Uint8Array(stringToArrayBuffer('a=1')));
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support DataView as body', () => {
-		const res = new Response(new DataView(stringToArrayBuffer('a=1')));
-		return res.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should default to null as body', () => {
-		const res = new Response();
-		expect(res.body).to.equal(null);
-
-		return res.text().then(result => expect(result).to.equal(''));
-	});
-
-	it('should default to 200 as status code', () => {
-		const res = new Response(null);
-		expect(res.status).to.equal(200);
-	});
-
-	it('should default to empty string as url', () => {
-		const res = new Response();
-		expect(res.url).to.equal('');
-	});
-});
-
-describe('Request', () => {
-	it('should have attributes conforming to Web IDL', () => {
-		const req = new Request('https://github.com/');
-		const enumerableProperties = [];
-		for (const property in req) {
-			enumerableProperties.push(property);
-		}
-
-		for (const toCheck of [
-			'body',
-			'bodyUsed',
-			'arrayBuffer',
-			'blob',
-			'json',
-			'text',
-			'method',
-			'url',
-			'headers',
-			'redirect',
-			'clone',
-			'signal'
-		]) {
-			expect(enumerableProperties).to.contain(toCheck);
-		}
-
-		for (const toCheck of [
-			'body', 'bodyUsed', 'method', 'url', 'headers', 'redirect', 'signal'
-		]) {
-			expect(() => {
-				req[toCheck] = 'abc';
-			}).to.throw();
-		}
-	});
-
-	it('should support wrapping Request instance', () => {
-		const url = `${base}hello`;
-
-		const form = new FormData();
-		form.append('a', '1');
-		const {signal} = new AbortController();
-
-		const r1 = new Request(url, {
-			method: 'POST',
-			follow: 1,
-			body: form,
-			signal
-		});
-		const r2 = new Request(r1, {
-			follow: 2
-		});
-
-		expect(r2.url).to.equal(url);
-		expect(r2.method).to.equal('POST');
-		expect(r2.signal).to.equal(signal);
-		// Note that we didn't clone the body
-		expect(r2.body).to.equal(form);
-		expect(r1.follow).to.equal(1);
-		expect(r2.follow).to.equal(2);
-		expect(r1.counter).to.equal(0);
-		expect(r2.counter).to.equal(0);
-	});
-
-	it('should override signal on derived Request instances', () => {
-		const parentAbortController = new AbortController();
-		const derivedAbortController = new AbortController();
-		const parentRequest = new Request(`${base}/hello`, {
-			signal: parentAbortController.signal
-		});
-		const derivedRequest = new Request(parentRequest, {
-			signal: derivedAbortController.signal
-		});
-		expect(parentRequest.signal).to.equal(parentAbortController.signal);
-		expect(derivedRequest.signal).to.equal(derivedAbortController.signal);
-	});
-
-	it('should allow removing signal on derived Request instances', () => {
-		const parentAbortController = new AbortController();
-		const parentRequest = new Request(`${base}/hello`, {
-			signal: parentAbortController.signal
-		});
-		const derivedRequest = new Request(parentRequest, {
-			signal: null
-		});
-		expect(parentRequest.signal).to.equal(parentAbortController.signal);
-		expect(derivedRequest.signal).to.equal(null);
-	});
-
-	it('should throw error with GET/HEAD requests with body', () => {
-		expect(() => new Request('.', {body: ''}))
-			.to.throw(TypeError);
-		expect(() => new Request('.', {body: 'a'}))
-			.to.throw(TypeError);
-		expect(() => new Request('.', {body: '', method: 'HEAD'}))
-			.to.throw(TypeError);
-		expect(() => new Request('.', {body: 'a', method: 'HEAD'}))
-			.to.throw(TypeError);
-		expect(() => new Request('.', {body: 'a', method: 'get'}))
-			.to.throw(TypeError);
-		expect(() => new Request('.', {body: 'a', method: 'head'}))
-			.to.throw(TypeError);
-	});
-
-	it('should default to null as body', () => {
-		const req = new Request(base);
-		expect(req.body).to.equal(null);
-		return req.text().then(result => expect(result).to.equal(''));
-	});
-
-	it('should support parsing headers', () => {
-		const url = base;
-		const req = new Request(url, {
-			headers: {
-				a: '1'
-			}
-		});
-		expect(req.url).to.equal(url);
-		expect(req.headers.get('a')).to.equal('1');
-	});
-
-	it('should support arrayBuffer() method', () => {
-		const url = base;
-		const req = new Request(url, {
-			method: 'POST',
-			body: 'a=1'
-		});
-		expect(req.url).to.equal(url);
-		return req.arrayBuffer().then(result => {
-			expect(result).to.be.an.instanceOf(ArrayBuffer);
-			const str = String.fromCharCode.apply(null, new Uint8Array(result));
-			expect(str).to.equal('a=1');
-		});
-	});
-
-	it('should support text() method', () => {
-		const url = base;
-		const req = new Request(url, {
-			method: 'POST',
-			body: 'a=1'
-		});
-		expect(req.url).to.equal(url);
-		return req.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support json() method', () => {
-		const url = base;
-		const req = new Request(url, {
-			method: 'POST',
-			body: '{"a":1}'
-		});
-		expect(req.url).to.equal(url);
-		return req.json().then(result => {
-			expect(result.a).to.equal(1);
-		});
-	});
-
-	it('should support buffer() method', () => {
-		const url = base;
-		const req = new Request(url, {
-			method: 'POST',
-			body: 'a=1'
-		});
-		expect(req.url).to.equal(url);
-		return req.buffer().then(result => {
-			expect(result.toString()).to.equal('a=1');
-		});
-	});
-
-	it('should support blob() method', () => {
-		const url = base;
-		const req = new Request(url, {
-			method: 'POST',
-			body: Buffer.from('a=1')
-		});
-		expect(req.url).to.equal(url);
-		return req.blob().then(result => {
-			expect(result).to.be.an.instanceOf(Blob);
-			expect(result.size).to.equal(3);
-			expect(result.type).to.equal('');
-		});
-	});
-
-	it('should support clone() method', () => {
-		const url = base;
-		let body = resumer().queue('a=1').end();
-		body = body.pipe(new stream.PassThrough());
-		const agent = new http.Agent();
-		const {signal} = new AbortController();
-		const req = new Request(url, {
-			body,
-			method: 'POST',
-			redirect: 'manual',
-			headers: {
-				b: '2'
-			},
-			follow: 3,
-			compress: false,
-			agent,
-			signal
-		});
-		const cl = req.clone();
-		expect(cl.url).to.equal(url);
-		expect(cl.method).to.equal('POST');
-		expect(cl.redirect).to.equal('manual');
-		expect(cl.headers.get('b')).to.equal('2');
-		expect(cl.follow).to.equal(3);
-		expect(cl.compress).to.equal(false);
-		expect(cl.method).to.equal('POST');
-		expect(cl.counter).to.equal(0);
-		expect(cl.agent).to.equal(agent);
-		expect(cl.signal).to.equal(signal);
-		// Clone body shouldn't be the same body
-		expect(cl.body).to.not.equal(body);
-		return Promise.all([cl.text(), req.text()]).then(results => {
-			expect(results[0]).to.equal('a=1');
-			expect(results[1]).to.equal('a=1');
-		});
-	});
-
-	it('should support ArrayBuffer as body', () => {
-		const req = new Request(base, {
-			method: 'POST',
-			body: stringToArrayBuffer('a=1')
-		});
-		return req.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support Uint8Array as body', () => {
-		const req = new Request(base, {
-			method: 'POST',
-			body: new Uint8Array(stringToArrayBuffer('a=1'))
-		});
-		return req.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-
-	it('should support DataView as body', () => {
-		const req = new Request(base, {
-			method: 'POST',
-			body: new DataView(stringToArrayBuffer('a=1'))
-		});
-		return req.text().then(result => {
-			expect(result).to.equal('a=1');
-		});
-	});
-});
-
-function streamToPromise(stream, dataHandler) {
-	return new Promise((resolve, reject) => {
-		stream.on('data', (...args) => {
-			Promise.resolve()
-				.then(() => dataHandler(...args))
-				.catch(reject);
-		});
-		stream.on('end', resolve);
-		stream.on('error', reject);
-	});
-}
-
-describe('external encoding', () => {
-	describe('data uri', () => {
-		it('should accept data uri', () => {
-			return fetch('data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs=').then(r => {
-				expect(r.status).to.equal(200);
-				expect(r.headers.get('Content-Type')).to.equal('image/gif');
-
-				return r.buffer().then(b => {
-					expect(b).to.be.an.instanceOf(Buffer);
-				});
-			});
-		});
-
-		it('should accept data uri of plain text', () => {
-			return fetch('data:,Hello%20World!').then(r => {
-				expect(r.status).to.equal(200);
-				expect(r.headers.get('Content-Type')).to.equal('text/plain');
-				return r.text().then(t => expect(t).to.equal('Hello World!'));
-			});
-		});
-
-		it('should reject invalid data uri', () => {
-			return fetch('data:@@@@').catch(e => {
-				expect(e).to.exist;
-				expect(e.message).to.include('invalid URL');
-			});
-		});
 	});
 });
